save conversations to db
