+++ src/server/mod.rs
@@ -1,0 +1,5 @@
+pub mod api;
+#[cfg(feature = "ssr")]
+pub mod db;
+#[cfg(feature = "ssr")]
+pub mod ollama;
